move files around for github diff
